refactor(timscentroid): move TimedStep down one layer
Relocate TimedStep + LABEL_WIDTH from timsseek::scoring::timings into
timscentroid so the upcoming tims_stage crate can reuse it without a
layer inversion. timsseek's scoring::timings keeps its scoring-specific
report types (ScoreTimings, PrescoreTimings, PipelineReport, etc.) and
re-exports TimedStep from timscentroid so existing caller paths
(timsseek::TimedStep, timsseek::scoring::TimedStep,
timsseek::scoring::timings::TimedStep) all continue to work.
